Add files via upload
The 2 files are used for inserting data automatically much faster than any human could ever put data in a database and the black wall for data verification and alteration. It also has a built in anomaly handler so thats a neat perk of it. Everytime its used in the  generation data anomalies get cleared first then they get inserted into the table.